Collect a `__GLIBC_MINOR__` definition from `zig cc` if any.
This is specific to Zig because while Zig ships with the recent enough
glibc it dynamically sets `__GLIBC_MINOR__` according to the target.

At this point it seems more robust to collect *all* options from `zig
cc` output, but this is not as viable because all we can see is cc1
options (`-###` or `-v`), not clang driver options from Zig. It seems
that we have no direct means to collect the latter, so we have to be
conservative about which options to collect.