Improve compatibility with older clang.
It turned out that unused argument warnings (`-Wunused-command-line-
argument`) come from cc1, not from driver, and as long as all options
are "claimed"---not necessarily used---it doesn't issue warnings.

On the other hands, Zig internally adds `-nostdinc` but we didn't
because we wanted header-only C++ libraries in the host to be usable,
but it turned out that the presence of `__has_include` made this
horribly fragile (e.g. some versions of libc++ used `__has_include(
<pthread.h>)` to determine pthread support), so we gave up. If this
facility is really desired, consider using `-idirafter /usr/include`.