fix: Handle -Wl,-exported_symbols_list <path>
When building a targeting a macos target you may get an error such as:

error: unsupported linker arg: -exported_symbols_list

A previous fix addressed the first variant of this flag where it is one argument.
This fix addresses the second form of this flag where the symbols list is passed
as a file path via a follow up argument.

Repros by building a cdylib targetting aarch64-apple-darwin.

Fixes #164