Don't remove temporary empty files from the cache.
When multiple processes try to build with Cargo-ZigBuild in parallel,
the processes compete writing and removing these files, which
can cause the builds to fail.

Since these empty files don't seem problematic, the most simple
solution is to keep the files in the cache and reuse them for
all invocations. That removes the race condition when two builders
try to use them in parallel.